% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scrape_colors.R
\name{download_images}
\alias{download_images}
\title{a function for downloading images from http://wesandersonpalettes.tumblr.com/}
\usage{
download_images(image_urls, folder = "")
}
\description{
a function for downloading images from http://wesandersonpalettes.tumblr.com/
}

