% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scrape_colors.R
\name{get_image_links}
\alias{get_image_links}
\title{a function for getting images links from http://wesandersonpalettes.tumblr.com/}
\usage{
get_image_links(url = "http://wesandersonpalettes.tumblr.com/")
}
\description{
a function for getting images links from http://wesandersonpalettes.tumblr.com/
}

