% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scrape_colors.R
\name{get_wes_andersons}
\alias{get_wes_andersons}
\title{a function to get wes anderson color schemes from the web}
\usage{
get_wes_andersons()
}
\description{
a function to get wes anderson color schemes from the web
}

