% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scrape_colors.R
\name{process_images}
\alias{process_images}
\title{a function for processing images and extracitng colors}
\usage{
process_images(images)
}
\description{
a function for processing images and extracitng colors
}

